+++ resources/lang/en/notifications.php
@@ -20,4 +20,5 @@ return [
     'password-reset.invalid_token'    => 'Reset token is invalid',
     'password-reset.password_updated' => 'Password updated',
     'password-reset.email_sent'       => 'Password reset sent',
+    'customer.updated'                => 'Customer Updated',
 ];