+++ routes/includes/customers.php
@@ -1,5 +1,6 @@
 <?php
 
+use GetCandy\Hub\Http\Livewire\Pages\Customers\CustomerShow;
 use GetCandy\Hub\Http\Livewire\Pages\Customers\CustomersIndex;
 use Illuminate\Support\Facades\Route;
 
@@ -7,7 +8,13 @@ use Illuminate\Support\Facades\Route;
  * Channel routes.
  */
 Route::group([
-    // 'middleware' => 'can:catalogue:manage-customers',
+    'middleware' => 'can:catalogue:manage-customers',
 ], function () {
     Route::get('/', CustomersIndex::class)->name('hub.customers.index');
+
+    Route::group([
+        'prefix' => '{customer}',
+    ], function () {
+        Route::get('/', CustomerShow::class)->name('hub.customers.show');
+    });
 });