+++ src/AdminHubServiceProvider.php
@@ -11,6 +11,8 @@ use GetCandy\Hub\Http\Livewire\Components\Collections\CollectionGroupShow;
 use GetCandy\Hub\Http\Livewire\Components\Collections\CollectionGroupsIndex;
 use GetCandy\Hub\Http\Livewire\Components\Collections\CollectionShow;
 use GetCandy\Hub\Http\Livewire\Components\Collections\SideMenu;
+use GetCandy\Hub\Http\Livewire\Components\Customers\CustomerShow;
+use GetCandy\Hub\Http\Livewire\Components\Customers\CustomersIndex;
 use GetCandy\Hub\Http\Livewire\Components\Orders\OrderShow;
 use GetCandy\Hub\Http\Livewire\Components\Orders\OrdersIndex;
 use GetCandy\Hub\Http\Livewire\Components\ProductOptions\OptionManager;
@@ -65,7 +67,7 @@ use Livewire\Livewire;
 
 class AdminHubServiceProvider extends ServiceProvider
 {
-    protected $configFiles = ['products'];
+    protected $configFiles = ['products', 'customers'];
 
     protected $root = __DIR__.'/..';
 
@@ -154,6 +156,7 @@ class AdminHubServiceProvider extends ServiceProvider
         $this->registerReportingComponents();
         $this->registerSettingsComponents();
         $this->registerOrderComponents();
+        $this->registerCustomerComponents();
 
         // Blade Components
         Blade::componentNamespace('GetCandy\\Hub\\Views\\Components', 'hub');
@@ -188,6 +191,12 @@ class AdminHubServiceProvider extends ServiceProvider
     {
         Livewire::component('hub.components.orders.index', OrdersIndex::class);
         Livewire::component('hub.components.orders.show', OrderShow::class);
+    }
+
+    protected function registerCustomerComponents()
+    {
+        Livewire::component('hub.components.customers.index', CustomersIndex::class);
+        Livewire::component('hub.components.customers.show', CustomerShow::class);
     }
 
     /**